Adding suggestion for IGP's coding villa!
I'm not sure if the brackets match up, this is not really necessary but why not waste my time rather than doing nothing with is. If this p-r gets merged.. make sure to fill the channel id in line 19.. I did this within 5 minutes so idk if its all right or the brackets arent gonna match lol.